requirements.txt
newsapi-python
bs4
python-dotenv